Bugfix : 스타가 팬미팅 신청 LocalDateTime 오류 수정

See merge request s06-webmobile1-sub2/S06P12E204!15

+++ Backend/src/main/java/com/ssafy/yourstar/domain/meeting/service/MeetingServiceImpl.java
@@ -64,9 +64,13 @@ public class MeetingServiceImpl implements MeetingService {
     }
 
     @Override
-    public boolean meetingRemoveByUser(Applicant applicant) {
-        // 해당 팬미팅이 존재하는지 조회 후 있을 때 삭제하는 코드로 수정
-        applicantRepository.delete(applicant);
-        return true;
+    public boolean meetingRemoveByUser(int memberId, int meetingId) {
+        // 해당 팬미팅이 존재하는지 조회 후 있을 때 삭제
+        if (applicantRepository.findByMemberIdAndMeetingId(memberId, meetingId).isPresent()) {
+            Applicant applicant = applicantRepository.findByMemberIdAndMeetingId(memberId, meetingId).get();
+            applicantRepository.delete(applicant);
+            return true;
+        }
+        return false;
     }
 }
